requirements.txt
httpx
beautifulsoup4
feedparser
pyyaml
python-dotenv
curl_cffi
playwright
streamlit
